requirements.txt
Flask
google-generativeai
pytesseract
PyMuPDF
Pillow
reportlab
